back/requirements.txt
django==4.2.0
asgiref
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
PyJWT
pytz
sqlparse
numpy
pandas
matplotlib
Pillow
social-auth-app-django
djoser
django_cleanup
psutil
docker
python-decouple
django-allauth
drf_social_oauth2
google-auth
httpie
google-api-python-client
mysqlclient
python-dotenv
gunicorn
